fix offenes
