--- Teacher.rb
@@ -1,11 +1,0 @@
-require './Person'
-
-class Teacher < Person
-    def initialize(specialization, name = "Unknown", age, parent_permission : true)
-    super (name, age, parent_permission)
-    @specialization = specialization
-    end
-    def can_use_services?
-        true
-    end
-end
